% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geokodigo.R
\docType{package}
\name{geokodiog}
\alias{geokodigo}
\alias{geokodigo-package}
\alias{geokodiog}
\title{An Interface to the Philippine Standard Geographical Codes}
\description{
The Philippine Standard Geographic Code (PSGC) is a classification and coding
of the geographical-political subdivisions of the country, such as the
region, the province, the municipality/city and the barangay. The PSGC is
currently compiled and updated quarterly through a collaborative effort of
the Philippine Statistics Authority (PSA) and its interagency Technical
Working Group on Geographic Code (TWGGC) composed of the Department of the
Interior and Local Government (DILG), Commission on Elections (COMELEC),
National Mapping and Resource Information Authority (NAMRIA), and the Land
Management Bureau (LMB). This package provides utilities for accessing
different versions of the PSGC over time, formatting and structuring them
into machine-readable relational database-ready formats.
}
\author{
\strong{Maintainer}: Ernest Guevarra \email{ernest@guevarra.io} (\href{https://orcid.org/0000-0002-4887-4415}{ORCID})

}
\keyword{internal}
